Rewrite all material including README and examples
This commit represents a complete rewrite of the guide.
The main catalyst for this rewrite was that the material was outdated.
Specifically, the guide made reference to boostrap.yml which is no
longer applicable.

When rewriting this guide, the format of the code example was changed.
The complete and initial folders were not carried forward, as it
felt cluttered having 3 applications repeated 2 times. Additionally,
for the same reasons, only gradle is included as the build tool.
Since the commands are basic, the user should be able to copy paste
with limited trouble.

An additional client application was added to demonstrate how two
services can interact with one another by first communicating with
the service registry. This seems like a more real world example to me
than calling an endpoint and receiving JSON, as was described in
the original guide.