Use Maven as the build system
This commit favors Maven over Gradle as the preferred build
system for this guide. The existing Gradle files are left in
place as it may be beneficial to keep. However, the README
instructions are updated so that the user can follow along
with a single build tool.

This commit also introduces the maven wrapper to the project.

Maven was preferred over Gradle for this guide as it is the
build tool used by the underling project, Spring Cloud Netflix.